fix(openclaw): extract serial flush-plan ingest for line ratchet
src/openclaw-flush-plan-lifecycle.ts is grandfathered at 1573 lines. Move serial one-turn ingest and its counter helpers to a sibling module so the file stays under that ceiling.
